refactor: split views and models into separate files
