+++ library_system/library/viewsets.py
@@ -1,7 +1,8 @@
 from rest_framework import viewsets
 
-from .models import Author, Book
-from .serializers import AuthorSerializer, BookSerializer
+from library.models.author import Author
+from library.models.book import Book
+from library.serializers import AuthorSerializer, BookSerializer
 
 
 class AuthorViewSet(viewsets.ModelViewSet):
